Admid organizing, its taking shape..ish
