+++ spec/spec_helper.rb
@@ -3,3 +3,7 @@ SimpleCov.start
 
 require 'pry'
 require 'csv'
+require './lib/game.rb'
+require './lib/league.rb'
+require './lib/team.rb'
+require './lib/season.rb'
